Add cast

+++ test/common.hpp
@@ -410,7 +410,7 @@ namespace rocwmma
         else if(isNaN)
         {
             retval             = false;
-            max_relative_error = std::numeric_limits<TypeA>::signaling_NaN();
+            max_relative_error = double(std::numeric_limits<TypeA>::signaling_NaN());
         }
         else if(max_relative_error > (eps * tolerance))
         {
@@ -499,7 +499,7 @@ namespace rocwmma
         else if(isNaN)
         {
             retval             = false;
-            max_relative_error = std::numeric_limits<TypeA>::signaling_NaN();
+            max_relative_error = double(std::numeric_limits<TypeA>::signaling_NaN());
         }
         else if(max_relative_error > (eps * tolerance))
         {
@@ -622,7 +622,7 @@ namespace rocwmma
         if(isNaN)
         {
             retval           = false;
-            maxRelativeError = std::numeric_limits<TypeA>::signaling_NaN();
+            maxRelativeError = double(std::numeric_limits<TypeA>::signaling_NaN());
         }
         else if(maxRelativeError > (eps * tolerance))
         {
@@ -703,7 +703,7 @@ namespace rocwmma
         if(isNaN)
         {
             retval           = false;
-            maxRelativeError = std::numeric_limits<TypeA>::signaling_NaN();
+            maxRelativeError = double(std::numeric_limits<TypeA>::signaling_NaN());
         }
         else if(maxRelativeError > (eps * tolerance))
         {
